pyproject.toml
[project]
name = "investor-paradise"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk[web]",
    "google-genai",
    "pandas",
    "python-dotenv",
    "fastparquet",
    "certifi",
    "sqlalchemy>=2.0.44",
    "aiosqlite>=0.21.0",
    "rich>=14.2.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.black]
line-length = 88
target-version = ["py311"]
